requirements.txt
pymongo==3.10.1
sqlalchemy==1.3.16
psycopg2-binary==2.8.5
python-dotenv==0.12.0
vaderSentiment==3.3.1
tweepy==3.8.0
slackclient==2.5.0
pandas==1.0.3
